Add dashboard styles to support editor view
This adds styles to embedded dashboards to hide the lock icon that
displays when the user has the Editor role, as well as hiding the flyout
link above reports.  In order to support embedding Analytical Designer,
users must be in the Editor role.  When they are in the Editor role,
they see an undesirable link to "View this report" in that flyout that
takes them to the report editor view.