Fix issue with browser history
When updating the iFrame URL, it causes browsers to add the page a
second time into the browser history. When a number reports are embedded
on the page, updating a filter across them can add lots of false entries
into the browser history.  This change will remove the iFrame and add it
back with the new URL rather than navigate the URL of the existing
iFrame.